Add conversions of boards and games to gray scale images
I am attempting to to use deep convalutional learning for the neural nets,
and images are the best format for that so far known.
Thus, this is allowing the conversion of all the saved sgf files into images
useable with those NNs.

The methods are split in two parts: gray scaling a board and gray scaling a
sequence of moves representing a game.
The Position object is suited to this as it handles all the internal tracking
of making a move, capturing stones, disallowing illegal play etc.