Add support for handicap stones
These stones were missing from the parsing of sgf files, and this caused some
very odd exceptions to be raised.
These changes are some of the additional error messges I added to help narrow
this down, the test I added to catch the problem I discovered as well as the
fix I made to solve the issue on the thick goban side.

Next I have to add support to the sgf parsing in the gofamiliar.