Update tests to use setup instead of handicap
The Position object was updated to call for setup, as the sgfs can
require more pre-game setup then just handicap stones.
As such, the name was changed for the keyword argument.
This also included a shift to forcing move colour for setup stones, and the
tests had to be updated also.