Add argument to disallow preventing playing in friendly eyes
I blocked such plays for the random playouts, as it would indiscriminantly
play in groups and self kill too much.
But it is not something to disallow for an intelligent player.
So I've added an argument to turn that on or off as a feature of the move
method.
The argument had to be pulled through the tests as well.