Do not bring groups back to life to find their size
I was using the pop method to get the single item out of a set.
This is the most pythonic way to get at that item, but it also removes the
item, which is not correct, as the dead stones need to be removed later on
in the move processing, ie they need to stay unpopped from the set.
I instead used the less nice list conversion and 0 index extraction method.